新增公告管理，友链管理

+++ src/main/java/com/pibigstar/repository/AnnouncementRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.pibigstar.domain.Announcement;
 
-public interface AnnouncementRepository extends JpaRepository<Announcement, Integer>{
+public interface AnnouncementRepository extends JpaRepository<Announcement, Long>{
 
 }
